moved middleware to sub-package
