--- src/api/appSessionMiddleware.js
@@ -1,5 +1,0 @@
-export default function uowMiddleware(req, res, next) {
-    req.appSession = {};
-
-    return next();
-}
