+++ src/api/middleware/appSession.js
@@ -1,0 +1,4 @@
+export default function (req, res, next) {
+    req.appSession = {};
+    return next();
+}
